補齊前三週進度，增加 can 的middleware，安裝npm、讓新增表單、儲存資料、刪除資料路由只能讓admin使用，自動生成三個 users: admin, manager, and user (方便我們做測試之用)、讓每一個路由都必須經過 auth 的 middleware (即必須登入帳號才可以使用)。、增加 can 的middleware，讓新增表單、儲存資料、刪除資料路由只能讓admin使用、不允許編輯表單給 user 權限使用，但允許 admin或manager使用

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ProductsController;
 use App\Http\Controllers\StoresController;
+use Illuminate\Support\Facades\Auth;
 
 /*
 |--------------------------------------------------------------------------
@@ -15,26 +16,61 @@ use App\Http\Controllers\StoresController;
 |
 */
 
-Route::get('/', function() {
-    return redirect('products');
+Route::middleware(['auth'])->group(function () {
+// 註解一下
+    Route::get('/', function() {
+        return redirect('products');
+    });
+
+    // 顯示顯示所有商品資料
+    Route::get('products', [ProductsController::class, 'index'])->name('products.index');
+    // 顯示單一商品資料
+    Route::get('products/{id}', [ProductsController::class, 'show'])->where('id', '[0-9]+')->name('products.show');
+    // 選定位置查詢商品
+    Route::get('products/discount', [ProductsController::class, 'discount'])->name('products.discount');
+    // 修改單一商品表單
+    Route::get('products/{id}/edit', [ProductsController::class, 'edit'])->where('id', '[0-9]+')->name('products.edit');
+    // 刪除單一商品資料
+    Route::delete('products/delete/{id}', [ProductsController::class, 'destroy'])->where('id', '[0-9]+')->name('products.destroy')->middleware('can:admin');
+    // 新增商品表單
+    Route::get('products/create', [ProductsController::class, 'create'])->name('products.create')->middleware('can:admin');
+    // 修改商品表單
+    Route::get('products/{id}/edit', [ProductsController::class, 'edit'])->where('id', '[0-9]+')->name('products.edit');
+    // 修改商品資料
+    Route::patch('products/update/{id}', [ProductsController::class, 'update'])->where('id', '[0-9]+')->name('products.update');
+    // 儲存新商品資料
+    Route::post('products/store', [ProductsController::class, 'store'])->where('id', '[0-9]+')->name('products.store')->middleware('can:admin');
+
+
+    // 顯示顯示所有商店資料
+    Route::get('stores', [StoresController::class, 'index'])->name('stores.index');
+    // 顯示單一商店資料
+    Route::get('stores/{id}', [StoresController::class, 'show'])->where('id', '[0-9]+')->name('stores.show');
+    // 修改單一商店表單
+    Route::get('stores/{id}/edit', [StoresController::class, 'edit'])->where('id', '[0-9]+')->name('stores.edit');
+    // 刪除單一商店及旗下商品資料
+    Route::delete('stores/delete/{id}', [StoresController::class, 'destroy'])->where('id', '[0-9]+')->name('stores.destroy')->middleware('can:admin');
+    // 新增商店表單
+    Route::get('stores/create', [StoresController::class, 'create'])->name('stores.create')->middleware('can:admin');
+    // 修改商店表單
+    Route::get('stores/update/{id}', [StoresController::class, 'update'])->where('id', '[0-9]+')->name('stores.update');
+    // 修改商店資料
+    Route::patch('stores/update/{id}', [StoresController::class, 'update'])->where('id', '[0-9]+')->name('stores.update');
+    // 儲存新商店資料
+    Route::post('stores/store', [StoresController::class, 'store'])->name('stores.store')->middleware('can:admin');
+    
+
 });
 
-// 顯示顯示所有商品資料
-Route::get('products', [ProductsController::class, 'index'])->name('products.index');
-// 顯示單一商品資料
-Route::get('products/{id}', [ProductsController::class, 'show'])->where('id', '[0-9]+')->name('products.show');
-// 修改單一商品表單
-Route::get('products/{id}/edit', [ProductsController::class, 'edit'])->where('id', '[0-9]+')->name('products.edit');
-// 刪除單一商品資料
-Route::delete('products/delete/{id}', [productsController::class, 'destroy'])->where('id', '[0-9]+')->name('products.destroy');
 
 
-// 顯示顯示所有商店資料
-Route::get('stores', [StoresController::class, 'index'])->name('stores.index');
-// 顯示單一商店資料
-Route::get('stores/{id}', [StoresController::class, 'show'])->where('id', '[0-9]+')->name('stores.show');
-// 修改單一商店表單
-Route::get('stores/{id}/edit', [StoresController::class, 'edit'])->where('id', '[0-9]+')->name('stores.edit');
-// 刪除單一商店及旗下商品資料
-Route::delete('stores/delete/{id}', [StoresController::class, 'destroy'])->where('id', '[0-9]+')->name('stores.destroy');
 
+
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
